# openapiv2/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "aruna/api/internal/v1/proxy.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/endpoint": {
      "get": {
        "summary": "Gets an specific endpoint by ID or Name",
        "operationId": "EndpointService_GetEndpoint",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetEndpointResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "endpointName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "endpointId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "EndpointService"
        ]
      },
      "post": {
        "operationId": "EndpointService_AddEndpoint",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AddEndpointResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AddEndpointRequest"
            }
          }
        ],
        "tags": [
          "EndpointService"
        ]
      }
    },
    "/v1/endpoint/default": {
      "get": {
        "summary": "This request returns the default endpoint for the current server\nIt may produce different results depending on the used server",
        "operationId": "EndpointService_GetDefaultEndpoint",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetDefaultEndpointResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "EndpointService"
        ]
      }
    },
    "/v1/endpoint/{endpointId}": {
      "delete": {
        "summary": "Deletes a specific endpoint by id\nThis needs admin permissions",
        "operationId": "EndpointService_DeleteEndpoint",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteEndpointResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "endpointId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "EndpointService"
        ]
      }
    },
    "/v1/endpoints": {
      "get": {
        "summary": "Gets all available endpoints",
        "operationId": "EndpointService_GetEndpoints",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetEndpointsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "EndpointService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "v1AddEndpointRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "epType": {
          "$ref": "#/definitions/v1EndpointType"
        },
        "proxyHostname": {
          "type": "string"
        },
        "internalHostname": {
          "type": "string"
        },
        "documentationPath": {
          "type": "string"
        },
        "isPublic": {
          "type": "boolean"
        }
      }
    },
    "v1AddEndpointResponse": {
      "type": "object",
      "properties": {
        "endpoint": {
          "$ref": "#/definitions/v1Endpoint"
        }
      }
    },
...
    "v1DeleteEndpointResponse": {
      "type": "object"
    },
...
    "v1Endpoint": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "epType": {
          "$ref": "#/definitions/v1EndpointType"
        },
        "name": {
          "type": "string"
        },
        "proxyHostname": {
          "type": "string"
        },
        "internalHostname": {
          "type": "string"
        },
        "documentationPath": {
          "type": "string"
        },
        "isPublic": {
          "type": "boolean"
        },
        "isDefault": {
          "type": "boolean"
        }
      }
    },
    "v1EndpointType": {
      "type": "string",
      "enum": [
        "ENDPOINT_TYPE_UNSPECIFIED",
        "ENDPOINT_TYPE_S3",
        "ENDPOINT_TYPE_FILE"
      ],
      "default": "ENDPOINT_TYPE_UNSPECIFIED"
    },
...
    "v1GetDefaultEndpointResponse": {
      "type": "object",
      "properties": {
        "endpoint": {
          "$ref": "#/definitions/v1Endpoint"
        }
      }
    },
...
    "v1GetEndpointResponse": {
      "type": "object",
      "properties": {
        "endpoint": {
          "$ref": "#/definitions/v1Endpoint"
        }
      }
    },
    "v1GetEndpointsResponse": {
      "type": "object",
      "properties": {
        "endpoints": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Endpoint"
          }
        }
      }
    },
...
  }
}